Renamed the directories
