+++ magicctapipe/image/__init__.py
@@ -1,0 +1,19 @@
+from .image import (
+    get_num_islands_MAGIC,
+    clean_image_params,
+    eval_impact,
+    tailcuts_clean_lstchain,
+    )
+
+from .stereo import (
+    write_hillas,
+    check_write_stereo,
+    check_stereo,
+    write_stereo,
+    )
+
+__all__ = ["get_num_islands_MAGIC", "clean_image_params",
+           "eval_impact", "tailcuts_clean_lstchain",
+           "write_hillas", "check_write_stereo",
+           "check_stereo", "write_stereo"]
+
