+++ magicctapipe/image/cleaning.py
