+++ magicctapipe/image/image.py
