+++ magicctapipe/image/stereo.py
