+++ magicctapipe/reco/direction_utils.py
