+++ magicctapipe/reco/energy_utils.py
