+++ magicctapipe/reco/event_processing.py
