+++ magicctapipe/reco/global_utils.py
