--- magicctapipe/train/__init__.py
@@ -1,66 +1,0 @@
-from .classifier_utils import (
-    GetHist_classifier,
-    evaluate_performance_classifier,
-    get_weights_classifier,
-    print_par_imp_classifier,
-    load_init_data_classifier,
-    check_train_test_intersections_classifier,
-    )
-
-from .direction_utils import (
-    compute_separation_angle_direction,
-    )
-
-from .energy_utils import (
-    GetHist2D_energy,
-    evaluate_performance_energy,
-    plot_migmatrix,
-    )
-
-from .event_processing import (
-    RegressorClassifierBase,
-    EnergyRegressor,
-    HillasFeatureSelector,
-    EventFeatureSelector,
-    EventFeatureTargetSelector,
-    EventProcessor,
-    EnergyEstimator,
-    EnergyEstimatorPandas,
-    DirectionEstimatorPandas,
-    EventClassifierPandas,
-    DirectionStereoEstimatorPandas,
-    )
-
-from .global_utils import (
-    compute_event_weights,
-    get_weights_mc_dir_class,
-    check_train_test_intersections,
-    )
-
-__all__ = [
-    "GetHist_classifier",
-    "evaluate_performance_classifier",
-    "get_weights_classifier",
-    "print_par_imp_classifier",
-    "load_init_data_classifier",
-    "check_train_test_intersections_classifier",
-    "compute_separation_angle_direction",
-    "GetHist2D_energy",
-    "evaluate_performance_energy",
-    "plot_migmatrix",
-    "RegressorClassifierBase",
-    "EnergyRegressor",
-    "HillasFeatureSelector",
-    "EventFeatureSelector",
-    "EventFeatureTargetSelector",
-    "EventProcessor",
-    "EnergyEstimator",
-    "EnergyEstimatorPandas",
-    "DirectionEstimatorPandas",
-    "EventClassifierPandas",
-    "DirectionStereoEstimatorPandas",
-    "compute_event_weights",
-    "get_weights_mc_dir_class",
-    "check_train_test_intersections",
-]
-
